pyproject.toml
[tool.poetry]
name = "mahad"
version = "0.2.0"
description = "An Arabic text processing library intended for use in NLP applications."
authors = ["Mohammad Al-Fetyani <m4bh@hotmail.com>"]
license = "BSD-3-Clause"
readme="README.md"
repository="https://github.com/TRoboto/Maha"
documentation="https://maha.readthedocs.io"
keywords=["Arabic", "NLP", "Text Processing", "Parsers", "Stream Text Processing"]

classifiers= [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD License",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
    "Topic :: Utilities",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Natural Language :: Arabic",
    "Natural Language :: English",
    ]
exclude = ["tools/","images/",".vscode/"]
packages = [
    { include = "maha" },
]

[tool.poetry.dependencies]
python = "^3.7.1"
tqdm = "^4.61.1"
python-dateutil = "^2.8.2"
regex = "^2021.8.28"
typing-extensions = "^3.10.0"
hijri-converter = "^2.2.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
black = "^21.5b1"
isort = "^5.8.0"
Sphinx = "^4.1.2"
furo = "^2021.8.31"
pytest = "^6.2.4"
sphinx-copybutton = "^0.3.1"
pytest-cov = "^2.12.1"
blacken-docs = "^1.10.0"
mypy = "^0.910"
tox = "^3.24.3"
types-python-dateutil = "^2.8.0"
sphinx-autoapi = "^1.8.4"
GitPython = "^3.1.24"
PyGithub = "^1.55"
linuxdoc = "^20210324"
datasets = "^1.18.2"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/TRoboto/Maha/issues"
"Discord" = "https://discord.gg/6W2tRFE7k4"
"Documentation" = "https://maha.readthedocs.io"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
exclude = "tests|tools|docs"
ignore_missing_imports = true


[build-system]
requires = ["setuptools","poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
